More Link is pointing to the custom URL
In views, when one wants to provide "more" link after certain number of items, there should be a block and the link is pointing to the page. But when this page should be built with panels, this approach is not working, because the panel page would have the same URL as the view page. That's why more link is pointing to the custom URL, which is a page built with panels, one of the panels is view of a type content pane with complete list of items...

Signed-off-by: Kresimir Nikolic <kresimir.nikolic@gmail.com>