menu_tree Theme Function extended: added menu options "divider" and "dropdown-header"
To add divider: add link in main menu, 	call it DIVIDER, insert "http://#divider" in path field, and drag the divider to desired position.

To add heading: add link in main menu, call it HEADING name, insert "http://#dropdown-header" in path field, add the name to the description field, and drag the heading to desired position.

Signed-off-by: Kresimir Nikolic <kresimir.nikolic@gmail.com>