Module updated: Panels Bootstrap Layouts
Because adding a custom bootstrap layout (in /modules/custom/iaea_bootstrap_layouts) was not working, the original layouts were tweaked: hero region has been added.

Also, because hero div has to be outside of the rest of the content, jQuery is used to manipulate the DOM, .hero is removed from .container and added directly under body, one step higher.

Signed-off-by: Kresimir Nikolic <kresimir.nikolic@gmail.com>